remove useless files
